Resolve memory leak issue (feat. jit)
Memory leak issue came from logging values as well as inside of value_and_grad function.
The jax internal issue can be resolved if jit wraps the function.
But, the compilation time is so long that we can't do a quick test. Thus, --no-jit option is added to disable jit.